+++ cf2cdm/cfcoords.py
@@ -175,6 +175,13 @@ def translate_coords(
         except:
             if errors != 'ignore':
                 raise RuntimeError("error while translating coordinate: %r" % cf_name)
+    config = coord_model.get('config', {})
+    if config.get('preferred_time_dimension', 'time') == 'valid_time':
+        try:
+            data = ensure_valid_time(data)
+        except Exception:
+            if errors != 'ignore':
+                raise RuntimeError("error while ensuring valid_time coordinate")
     return data
 
 
@@ -183,14 +190,13 @@ def ensure_valid_time_present(data, valid_time_name='valid_time'):
     valid_times = match_values(is_valid_time, data.coords)
     times = match_values(is_time, data.coords)
     steps = match_values(is_step, data.coords)
-    time = step = ''
+    time = times[0] if times else ''
+    step = steps[0] if steps else ''
     if not valid_times:
-        if not times:
+        if not time:
             raise ValueError("not enough information to ensure a 'valid_time'.")
         valid_time = valid_time_name
-        time = times[0]
-        if steps:
-            step = steps[0]
+        if step:
             data.coords[valid_time] = data.coords[time] + data.coords[step]
         else:
             data.coords[valid_time] = data.coords[time]
@@ -204,8 +210,13 @@ def ensure_valid_time(data):
     # type: (xr.Dataset) -> xr.Dataset
     valid_time, time, step = ensure_valid_time_present(data)
     if valid_time not in data.dims:
-        if data.coords[time].size == data.coords[valid_time].size:
+        if time and time in data.dims and data.coords[time].size == data.coords[valid_time].size:
             return data.swap_dims({time: valid_time})
-        if data.coords[step].size == data.coords[valid_time].size:
+        if step and step in data.dims and data.coords[step].size == data.coords[valid_time].size:
             return data.swap_dims({step: valid_time})
+        # also convert is valid_time can index all times and steps
+        if step and time and step in data.dims and time in data.dims and \
+                data.coords[step].size * data.coords[time].size == data.coords[valid_time].size:
+            data = data.stack(tmp_coord=(time, step))
+            data = data.swap_dims({'tmp_coord': valid_time}).drop('tmp_coord').dropna(valid_time)
     return data